Revert "Made optional the automatic addition of PLC status template"
This reverts commit ed75bfa75fe70cab7b7b7efb1f78e5c81750db97.

Existing IOCs updating to a dlsPLC version which depends on this ether_ip version 3-1dls1-1 lost the initialisation of the port and driver for the NX102 PLC from the bootscript.  This change to make something optional did not produce the correct behaviour for existing IOCs with the default of "True".  Since fix is needed urgently for operational IOCs and Armando is on holiday reverting this change.